+++ packages/util-crypto/src/address/setSS58Format.ts
@@ -6,6 +6,10 @@ import { Prefix } from './types';
 
 import defaults from './defaults';
 
+/**
+ * @description Sets the global SS58 format to use for address encoding
+ * @deprecated Use keyring.setSS58Format
+ */
 export default function setSS58Format (prefix: Prefix): void {
   defaults.prefix = prefix;
 }