+++ packages/util/src/format/formatBalance.ts
@@ -18,6 +18,7 @@ interface Defaults {
 }
 
 interface Options {
+  decimals?: number;
   forceUnit?: string;
   withSi?: boolean;
   withSiFull?: boolean;
@@ -40,7 +41,7 @@ let defaultDecimals = DEFAULT_DECIMALS;
 let defaultUnit = DEFAULT_UNIT;
 
 // Formats a string/number with <prefix>.<postfix><type> notation
-function _formatBalance <ExtToBn extends ToBn> (input?: number | string | BN | ExtToBn, options: Options | boolean = true, decimals: number = defaultDecimals): string {
+function _formatBalance <ExtToBn extends ToBn> (input?: number | string | BN | ExtToBn, options: Options | boolean = true, optDecimals: number = defaultDecimals): string {
   let text = bnToBn(input).toString();
 
   if (text.length === 0 || text === '0') {
@@ -56,7 +57,7 @@ function _formatBalance <ExtToBn extends ToBn> (input?: number | string | BN | E
   }
 
   // extract options - the boolean case is for backwards-compat
-  const { forceUnit = undefined, withSi = true, withSiFull = false, withUnit = true } = isBoolean(options)
+  const { decimals = optDecimals, forceUnit = undefined, withSi = true, withSiFull = false, withUnit = true } = isBoolean(options)
     ? { withSi: options }
     : options;