+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # 2.4.0-beta.x
 
-- Ensure that the `formatBalance` no unit overrides apply when no SI is applicable
-- Adjust `formatBalance` to take `withUnit: string | boolean` where string is an actual unit, e.g. `KSM`
+- Ensure that `formatBalance` does not apply unit overrides apply when no SI is applicable
+- Adjust `formatBalance` to take expanded `withUnit: string | boolean` option where string is an actual unit, e.g. `KSM`
+- The `.setSS58Format` on keyring will now just operate on the keyring, not globally. The global `setSS58Format` in `@polkadot/util-crypto` is deprecated.
+- The deprecated `addressPrefix` option to keyring has been belatedly removed (Use `ss58Format` rather)
 
 # 2.3.1 Feb 15, 2020