+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-# 2.4.0-beta.x
+# 2.4.1 Feb 18, 2020
 
 - Ensure that `formatBalance` does not apply unit overrides apply when no SI is applicable
-- Adjust `formatBalance` to take expanded `withUnit: string | boolean` option where string is an actual unit, e.g. `KSM`
-- The `.setSS58Format` on keyring will now just operate on the keyring, not globally. The global `setSS58Format` in `@polkadot/util-crypto` is deprecated.
+- Adjust `formatBalance(<balance>, <options>)` to take expanded `withUnit: string | boolean` option where string is an actual unit, e.g. `KSM`
+- The `decimals` option to `formatBalance` should now be passed as part of the `options` (previous last-param deprecated)
+- The `.setSS58Format` on keyring will now just operate on pairs created on the keyring, not globally. The global `setSS58Format` in `@polkadot/util-crypto` will be deprecated.
 - The deprecated `addressPrefix` option to keyring has been belatedly removed (Use `ss58Format` rather)
 
 # 2.3.1 Feb 15, 2020